+++ src/framework/piral-base/src/index.minimal.ts
@@ -1,0 +1,6 @@
+import 'systemjs/dist/system.js';
+import 'systemjs/dist/extras/named-register.js';
+
+export * from './events';
+export * from './types';
+export * from './utils/system';